Use journald logging for firecracker too
