+++ src/aleph/vm/controllers/firecracker/executable.py
@@ -188,6 +188,7 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType], AlephVmControllerIn
 
         self.fvm = MicroVM(
             vm_id=self.vm_id,
+            vm_hash=vm_hash,
             firecracker_bin_path=settings.FIRECRACKER_PATH,
             jailer_base_directory=settings.JAILER_BASE_DIR,
             use_jailer=settings.USE_JAILER,
@@ -259,9 +260,6 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType], AlephVmControllerIn
                 await self.tap_interface.delete()
             raise
 
-        if self.enable_console:
-            self.fvm.start_printing_logs()
-
         await self.wait_for_init()
         logger.debug(f"started fvm {self.vm_id}")
         await self.load_configuration()
@@ -331,18 +329,3 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType], AlephVmControllerIn
 
     async def create_snapshot(self) -> CompressedDiskVolumeSnapshot:
         raise NotImplementedError()
-
-    def get_log_queue(self) -> asyncio.Queue:
-        queue: asyncio.Queue = asyncio.Queue(maxsize=1000)
-        # Limit the number of queues per VM
-
-        if len(self.fvm.log_queues) > 20:
-            logger.warning("Too many log queues, dropping the oldest one")
-            self.fvm.log_queues.pop(0)
-        self.fvm.log_queues.append(queue)
-        return queue
-
-    def unregister_queue(self, queue: asyncio.Queue):
-        if queue in self.fvm.log_queues:
-            self.fvm.log_queues.remove(queue)
-        queue.empty()
