+++ src/aleph/vm/hypervisors/qemu/qemuvm.py
@@ -22,7 +22,6 @@ class QemuVM:
     mem_size_mb: int
     interface_name: str
     qemu_process = None
-    log_queues: list[asyncio.Queue]
 
     def __repr__(self) -> str:
         if self.qemu_process:
@@ -39,7 +38,6 @@ class QemuVM:
         self.vcpu_count = config.vcpu_count
         self.mem_size_mb = config.mem_size_mb
         self.interface_name = config.interface_name
-        self.log_queues = []
         self.vm_hash = vm_hash
 
     @property
